Separation of courier creation and courier authorization tests
